# openapi.yaml
openapi: 3.1.0
info:
  title: VoiceBlender API
  description: >
    VoiceBlender bridges SIP and WebRTC voice calls with multi-party audio
    mixing, real-time speech-to-text, text-to-speech, AI agent integration,
    recording, and webhook-based event delivery.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /legs:
    post:
      operationId: createLeg
      summary: Originate an outbound SIP call
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLegRequest'
      responses:
        '201':
          description: Leg created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leg'
        '400':
          description: Invalid JSON, bad SIP URI, unknown codec, or unsupported type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listLegs
      summary: List all active legs
      tags: [Legs]
      responses:
        '200':
          description: Array of legs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Leg'
  /legs/{id}:
    parameters:
      - $ref: '#/components/parameters/LegId'

    get:
      operationId: getLeg
      summary: Get a single leg
      tags: [Legs]
      responses:
        '200':
          description: Leg details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leg'
        '404':
          $ref: '#/components/responses/LegNotFound'

    delete:
      operationId: deleteLeg
      summary: Hang up a leg
      tags: [Legs]
      responses:
        '200':
          description: Leg hung up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: hung_up
        '404':
          $ref: '#/components/responses/LegNotFound'
  /legs/{id}/answer:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: answerLeg
      summary: Answer a ringing inbound SIP leg
      tags: [Legs]
      responses:
        '200':
          description: Answer initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: answering
        '400':
          description: Not a SIP inbound leg
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg is not in ringing state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /legs/{id}/play:
    parameters:
      - $ref: '#/components/parameters/LegId'
    post:
      operationId: playLeg
      summary: Start audio playback to a leg
      tags: [Legs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackRequest'
      responses:
        '200':
          description: Playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackResponse'
        '400':
          description: Invalid JSON or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/LegNotFound'
        '409':
          description: Leg has no audio writer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /legs/{id}/play/{playbackID}:
    parameters:
      - $ref: '#/components/parameters/LegId'
      - $ref: '#/components/parameters/PlaybackId'
    delete:
      operationId: stopPlayLeg
      summary: Stop audio playback on a leg
      tags: [Legs]
      responses:
        '200':
          description: Playback stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: stopped
        '404':
          description: No playback in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /rooms/{id}/play:
    parameters:
      - $ref: '#/components/parameters/RoomId'
    post:
      operationId: playRoom
      summary: Play audio to a room
      tags: [Rooms]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaybackRequest'
      responses:
        '200':
          description: Playback started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaybackResponse'
        '400':
          description: Invalid JSON or volume out of range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/RoomNotFound'
        '409':
          description: Room has no participants
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /rooms/{id}/play/{playbackID}:
    parameters:
      - $ref: '#/components/parameters/RoomId'
      - $ref: '#/components/parameters/PlaybackId'
    delete:
      operationId: stopPlayRoom
      summary: Stop room playback
      tags: [Rooms]
      responses:
        '200':
          description: Playback stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              example:
                status: stopped
        '404':
          description: No playback in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    LegId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Leg ID
    RoomId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Room ID
    PlaybackId:
      name: playbackID
      in: path
      required: true
      schema:
        type: string
      description: Playback ID
  responses:
    LegNotFound:
      description: Leg not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RoomNotFound:
      description: Room not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Leg:
      type: object
      properties:
        leg_id:
          type: string
          description: Unique leg identifier (UUID)
        type:
          type: string
          enum: [sip_inbound, sip_outbound, webrtc]
        state:
          type: string
          enum: [ringing, connected, hung_up]
        room_id:
          type: string
          description: Room ID if the leg is in a room, empty otherwise
      required: [leg_id, type, state]
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required: [error]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ── Request bodies ──────────────────────────────────────────────────
    CreateLegRequest:
      type: object
      properties:
        type:
          type: string
          enum: [sip]
          description: Leg type
        uri:
          type: string
          description: SIP URI to dial
        from:
          type: string
          description: 'Caller ID / SIP From header (URI or "display" <uri>)'
        privacy:
          type: string
          description: SIP Privacy header value (e.g. "id", "none")
        ring_timeout:
          type: integer
          description: Seconds to wait for answer; 0 = no timeout
          default: 0
        codecs:
          type: array
          items:
            type: string
            enum: [PCMU, PCMA, G722, opus]
          description: Codec preference order
      required: [type, uri]
...
    PlaybackRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          description: URL of the audio file
        mime_type:
          type: string
          description: MIME type (e.g. audio/wav)
        repeat:
          type: integer
          description: Number of times to repeat playback
          default: 0
        volume:
          type: integer
          description: Volume adjustment in dB (-8 to 8)
          minimum: -8
          maximum: 8
          default: 0
      required: [url, mime_type]
...
    PlaybackResponse:
      type: object
      properties:
        playback_id:
          type: string
          description: Unique playback identifier
        status:
          type: string
          example: playing
      required: [playback_id, status]
...